ensure we have pods/log permission
